refactor: 메서드명 변경, 패키지 변경
