+++ src/test/java/domain/PositionTest.java
@@ -1,8 +1,8 @@
 package domain;
 
-import chess.domain.File;
-import chess.domain.Position;
-import chess.domain.Rank;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -48,4 +48,5 @@ public class PositionTest {
         //then
         assertThat(isDiagonalSame).isFalse();
     }
+
 }
